# Reduce Implicit Array/Hash Allocations For Method Calls Involving Splats