--- a/insns.def
+++ b/insns.def
@@ -505,7 +505,9 @@ expandarray
     vm_expandarray(GET_CFP(), ary, num, (int)flag);
 }
 
-/* concat two arrays */
+/* concat two arrays, without modifying first array.
+ * attempts to convert both objects to arrays using to_a.
+ */
 DEFINE_INSN
 concatarray
 ()
@@ -514,6 +516,20 @@ concatarray
 // attr bool leaf = false; /* has rb_check_array_type() */
 {
     ary = vm_concat_array(ary1, ary2);
+}
+
+/* concat second array to first array.
+ * first argument must already be an array.
+ * attempts to convert second object to array using to_a.
+ */
+DEFINE_INSN
+concattoarray
+()
+(VALUE ary1, VALUE ary2)
+(VALUE ary)
+// attr bool leaf = false; /* has rb_check_array_type() */
+{
+    ary = vm_concat_to_array(ary1, ary2);
 }
 
 /* call to_a on array ary to splat */