--- a/insns.def
+++ b/insns.def
@@ -532,6 +532,18 @@ concattoarray
     ary = vm_concat_to_array(ary1, ary2);
 }
 
+/* push given number of objects to array directly before. */
+DEFINE_INSN
+pushtoarray
+(rb_num_t num)
+(...)
+(VALUE val)
+// attr rb_snum_t sp_inc = -(rb_snum_t)num;
+{
+    const VALUE *objp = STACK_ADDR_FROM_TOP(num);
+    val = rb_ary_cat(*(objp-1), objp, num);
+}
+
 /* call to_a on array ary to splat */
 DEFINE_INSN
 splatarray