--- a/test/ruby/test_ast.rb
+++ b/test/ruby/test_ast.rb
@@ -661,7 +661,7 @@ class TestAst < Test::Unit::TestCase
       node ? [node.children[-4], node.children[-2]&.children, node.children[-1]] : []
     end
 
-    assert_equal([:*, nil, :&], forwarding.call('...'))
+    assert_equal([:*, [:**], :&], forwarding.call('...'))
   end
 
   def test_ranges_numbered_parameter