--- a/vm_callinfo.h
+++ b/vm_callinfo.h
@@ -25,6 +25,7 @@ enum vm_call_flag_bits {
     VM_CALL_ZSUPER_bit,         // zsuper
     VM_CALL_OPT_SEND_bit,       // internal flag
     VM_CALL_KW_SPLAT_MUT_bit,   // kw splat hash can be modified (to avoid allocating a new one)
+    VM_CALL_ARGS_SPLAT_MUT_bit, // args splat can be modified (to avoid allocating a new one)
     VM_CALL__END
 };
 
@@ -40,6 +41,7 @@ enum vm_call_flag_bits {
 #define VM_CALL_ZSUPER          (0x01 << VM_CALL_ZSUPER_bit)
 #define VM_CALL_OPT_SEND        (0x01 << VM_CALL_OPT_SEND_bit)
 #define VM_CALL_KW_SPLAT_MUT    (0x01 << VM_CALL_KW_SPLAT_MUT_bit)
+#define VM_CALL_ARGS_SPLAT_MUT  (0x01 << VM_CALL_ARGS_SPLAT_MUT_bit)
 
 struct rb_callinfo_kwarg {
     int keyword_len;