@k0kubun 2023-12-15
I benchmarked this PR with headline benchmarks of yjit-bench:

## Interpreter

It speeds up psych-load and sequel, but it slows down activerecord. Is the slowdown intended?

```
before: ruby 3.3.0dev (2023-12-14T16:09:17Z master 247ce712fc) [x86_64-linux]
after: ruby 3.3.0dev (2023-12-15T00:48:09Z vm-call-args-splat.. b28e0be92e) [x86_64-linux]

--------------  -----------  ----------  ----------  ----------  -------------  ------------
bench           before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
activerecord    106.6        0.3         109.4       0.3         0.98           0.97
chunky-png      952.3        0.3         950.5       0.2         1.00           1.00
erubi-rails     2900.2       0.6         2891.2      0.6         1.01           1.00
hexapdf         3958.5       0.7         3947.6      0.5         1.00           1.00
liquid-c        106.1        0.4         106.1       0.4         1.00           1.00
liquid-compile  107.0        0.2         107.3       0.2         1.00           1.00
liquid-render   257.4        0.3         256.7       0.2         1.00           1.00
lobsters        1468.6       0.5         1474.4      0.5         1.09           1.00
mail            210.6        0.3         210.8       0.2         1.01           1.00
psych-load      3286.2       0.1         3191.2      0.1         1.03           1.03
railsbench      3369.0       0.1         3379.6      0.0         1.00           1.00
ruby-lsp        93.2         0.9         93.6        0.2         0.99           1.00
sequel          116.3        0.4         114.0       0.4         1.01           1.02
--------------  -----------  ----------  ----------  ----------  -------------  ------------
```

## YJIT

For your information, this is the impact on benchmarks with YJIT enabled.

```
before: ruby 3.3.0dev (2023-12-14T16:09:17Z master 247ce712fc) +YJIT [x86_64-linux]
after: ruby 3.3.0dev (2023-12-15T00:48:09Z vm-call-args-splat.. b28e0be92e) +YJIT [x86_64-linux]

--------------  -----------  ----------  ----------  ----------  -------------  ------------
bench           before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
activerecord    52.7         0.5         54.9        0.5         0.96           0.96
chunky-png      556.6        0.1         556.3       0.3         1.00           1.00
erubi-rails     1600.3       0.6         1620.2      0.5         0.98           0.99
hexapdf         2068.8       0.9         2126.6      0.9         0.99           0.97
liquid-c        71.8         0.6         72.1        0.6         0.99           1.00
liquid-compile  74.6         0.4         74.3        0.4         1.01           1.00
liquid-render   94.2         0.6         94.6        0.3         1.00           1.00
lobsters        947.4        1.6         963.8       1.0         1.03           0.98
mail            149.1        0.3         149.4       0.3         1.00           1.00
psych-load      1906.3       0.1         1939.8      0.1         0.98           0.98
railsbench      2037.4       0.1         2079.6      0.0         0.98           0.98
ruby-lsp        59.4         24.9        61.4        26.8        0.99           0.97
sequel          88.2         0.4         87.7        0.5         1.01           1.01
--------------  -----------  ----------  ----------  ----------  -------------  ------------
```