@jeremyevans 2023-12-15
> I benchmarked this PR with headline benchmarks of yjit-bench:
> 
> ## Interpreter
> It speeds up psych-load and sequel, but it slows down activerecord. Is the slowdown intended?

No.  I'm not sure why activerecord would be slower and nothing else would.  I'm surprised it speeds up sequel and psych-load as much as it did.

> ## YJIT
> For your information, this is the impact on benchmarks with YJIT enabled.

YJIT slowdown is somewhat expected until YJIT supports the concattoarray and pushtoarray VM instructions.  YJIT could probably also get a speedup using the new callinfo flags this adds to avoid allocations.