@maximecb 2023-12-15
@jeremyevans are you intending to merge this PR in for 3.3? I'm a wary of merging this just 10 days away from the release.

At Shopify we've been in the process of testing and fixing a lot of bugs. We've discovered and fixed a number of issues affecting both CRuby and YJIT. Many of which had to do with method calls.

IMO merging this just before the release means we don't have enough time to test. This is a fairly significant design change. It's also unfortunate that we didn't have time to discuss implications for YJIT or implement any kind of support. 

For those not intimately familiar with how Ruby splat calls get compiled, can you provide some YARV examples of how the bytecode might look before/after with the new instructions you want to introduce?