@tenderlove 2024-01-25
In the future, I think we should try to get tests passing rather than disabling them. If you need help getting the Prism tests to pass, please feel free to ping any of the Prism maintainers (me, @kddnewton, etc). We'd like Prism to be treated as a first class library embedded in CRuby.

Thanks.