@jeremyevans 2024-01-25
Agreed.  In this case, I did not know prism would break.  This PR was last rebased on January 8, and the prism CI that broke wasn't added until January 10.  I would have discovered this if I rebased again right before merging, but I didn't do that (I usually only rebase right before merging if the last rebase was over a month ago).